Refactor(app): QuickList dir분리
